Fix MySql deprecated types: FLOAT(M,D), DOUBLE(M,D) - propal

+++ htdocs/install/mysql/tables/llx_propaldet.sql
@@ -22,46 +22,46 @@
 create table llx_propaldet
 (
   rowid								integer AUTO_INCREMENT PRIMARY KEY,
-  fk_propal							integer		NOT NULL,
-  fk_parent_line					integer		NULL,
-  fk_product						integer		NULL,
-  label								varchar(255) DEFAULT NULL,
+  fk_propal							integer		    NOT NULL,
+  fk_parent_line					integer		    NULL,
+  fk_product						integer		    NULL,
+  label								varchar(255)    DEFAULT NULL,
   description						text,
-  fk_remise_except					integer      NULL,               -- Lien vers table des remises fixes
-  vat_src_code						varchar(10)  DEFAULT '',		 -- Vat code used as source of vat fields. Not strict foreign key here.
-  tva_tx							double(6,3)  DEFAULT 0, 	     -- Vat rate
-  localtax1_tx               		double(6,3)  DEFAULT 0,    		 -- localtax1 rate
-  localtax1_type			 		varchar(10)	  	 NULL, 			 -- localtax1 type
-  localtax2_tx               		double(6,3)  DEFAULT 0,    		 -- localtax2 rate
-  localtax2_type			 		varchar(10)	  	 NULL, 			 -- localtax2 type
-  qty								real,                            -- quantity
-  remise_percent					real         DEFAULT 0,          -- pourcentage de remise
-  remise							real         DEFAULT 0,          -- montant de la remise (obsolete)
-  price								real,                            -- prix final (obsolete)
-  subprice							double(24,8) DEFAULT 0,          -- prix unitaire article
-  total_ht							double(24,8) DEFAULT 0,          -- Total HT de la ligne toute quantite et incluant remise ligne et globale
-  total_tva							double(24,8) DEFAULT 0,          -- Total TVA de la ligne toute quantite et incluant remise ligne et globale
-  total_localtax1					double(24,8) DEFAULT 0,          -- Total localtax1
-  total_localtax2					double(24,8) DEFAULT 0,          -- Total localtax2
-  total_ttc							double(24,8) DEFAULT 0,          -- Total TTC de la ligne toute quantite et incluant remise ligne et globale
-  product_type						integer    DEFAULT 0,            -- 0 or 1. Value 9 may be used by some modules (amount of line may not be included into generated discount if value is 9).
-  date_start						datetime   DEFAULT NULL,         -- date debut si service
-  date_end							datetime   DEFAULT NULL,         -- date fin si service
-  info_bits							integer      DEFAULT 0,          -- TVA NPR ou non
+  fk_remise_except					integer         NULL,           -- Lien vers table des remises fixes
+  vat_src_code						varchar(10)     DEFAULT '',		-- Vat code used as source of vat fields. Not strict foreign key here.
+  tva_tx							numeric(6,3)    DEFAULT 0, 	    -- Vat rate
+  localtax1_tx               		numeric(6,3)    DEFAULT 0,      -- localtax1 rate
+  localtax1_type			 		varchar(10)	  	NULL, 			-- localtax1 type
+  localtax2_tx               		numeric(6,3)    DEFAULT 0,    	-- localtax2 rate
+  localtax2_type			 		varchar(10)	  	NULL, 			-- localtax2 type
+  qty								real,                           -- quantity
+  remise_percent					real            DEFAULT 0,      -- pourcentage de remise
+  remise							real            DEFAULT 0,      -- montant de la remise (obsolete)
+  price								real,                           -- prix final (obsolete)
+  subprice							numeric(24,8)   DEFAULT 0,      -- prix unitaire article
+  total_ht							numeric(24,8)   DEFAULT 0,      -- Total HT de la ligne toute quantite et incluant remise ligne et globale
+  total_tva							numeric(24,8)   DEFAULT 0,      -- Total TVA de la ligne toute quantite et incluant remise ligne et globale
+  total_localtax1					numeric(24,8)   DEFAULT 0,      -- Total localtax1
+  total_localtax2					numeric(24,8)   DEFAULT 0,      -- Total localtax2
+  total_ttc							numeric(24,8)   DEFAULT 0,      -- Total TTC de la ligne toute quantite et incluant remise ligne et globale
+  product_type						integer         DEFAULT 0,      -- 0 or 1. Value 9 may be used by some modules (amount of line may not be included into generated discount if value is 9).
+  date_start						datetime        DEFAULT NULL,   -- date debut si service
+  date_end							datetime        DEFAULT NULL,   -- date fin si service
+  info_bits							integer         DEFAULT 0,      -- TVA NPR ou non
   
-  buy_price_ht						double(24,8) DEFAULT 0,          -- buying price
-  fk_product_fournisseur_price		integer      DEFAULT NULL,       -- reference of supplier price when line was added (may be used to update buy_price_ht current price when future invoice will be created)
+  buy_price_ht						numeric(24,8)   DEFAULT 0,      -- buying price
+  fk_product_fournisseur_price		integer         DEFAULT NULL,   -- reference of supplier price when line was added (may be used to update buy_price_ht current price when future invoice will be created)
   
-  special_code						integer      DEFAULT 0,          -- code for special lines (may be 1=transport, 2=ecotax, 3=option, moduleid=...)
-  rang								integer      DEFAULT 0,          -- ordre affichage sur la propal
-  fk_unit           integer      DEFAULT NULL,           -- lien vers table des unités
+  special_code						integer         DEFAULT 0,      -- code for special lines (may be 1=transport, 2=ecotax, 3=option, moduleid=...)
+  rang								integer         DEFAULT 0,      -- ordre affichage sur la propal
+  fk_unit                           integer         DEFAULT NULL,   -- lien vers table des unités
   
-  fk_multicurrency			integer,
-  multicurrency_code			varchar(255),
-  multicurrency_subprice		double(24,8) DEFAULT 0,
-  multicurrency_total_ht		double(24,8) DEFAULT 0,
-  multicurrency_total_tva	double(24,8) DEFAULT 0,
-  multicurrency_total_ttc	double(24,8) DEFAULT 0
+  fk_multicurrency			        integer,
+  multicurrency_code			    varchar(255),
+  multicurrency_subprice		    numeric(24,8)   DEFAULT 0,
+  multicurrency_total_ht		    numeric(24,8)   DEFAULT 0,
+  multicurrency_total_tva	        numeric(24,8)   DEFAULT 0,
+  multicurrency_total_ttc	        numeric(24,8)   DEFAULT 0
 )ENGINE=innodb;
 
 --
